+++ example/a.ts
@@ -2,9 +2,11 @@ import { Elysia, error, t } from '../src'
 import { req } from '../test/utils'
 
 const app = new Elysia()
-	.get('/', ({ path, request }) => {
-		console.log(request)
-
-		return path
+	.get('/id/:id', ({ params: { id } }) => {
+		return id
+	}, {
+		response: t.Object({
+			a: t.String()
+		})
 	})
 	.listen(3000)