+++ src/adapter/bun/compose.ts
@@ -1,11 +1,25 @@
-import { error, InternalRoute, redirect, type AnyElysia } from '../..'
+import {
+	error,
+	redirect,
+	ELYSIA_TRACE,
+	type InternalRoute,
+	type AnyElysia
+} from '../..'
+
 import { createHoc, createOnRequestHandler } from '../../compose'
-import { sucrose } from '../../sucrose'
+import { sucrose, type Sucrose } from '../../sucrose'
+
+import { randomId, ELYSIA_REQUEST_ID } from '../../utils'
 
 const allocateIf = (value: string, condition: unknown) =>
 	condition ? value : ''
 
-const createContext = (app: AnyElysia, route: InternalRoute) => {
+const createContext = (
+	app: AnyElysia,
+	route: InternalRoute,
+	inference: Sucrose.Inference,
+	isInline = false
+) => {
 	let fnLiteral = ''
 
 	// @ts-expect-error private
@@ -20,14 +34,6 @@ const createContext = (app: AnyElysia, route: InternalRoute) => {
 
 	const standardHostname = app.config.handler?.standardHostname ?? true
 
-	const inference = sucrose(
-		Object.assign({}, route.hooks, {
-			handler: route.handler
-		}),
-		// @ts-expect-error
-		app.inference
-	)
-
 	const getQi =
 		`const u=request.url,` +
 		`s=u.indexOf('/',${standardHostname ? 11 : 7}),` +
@@ -35,19 +41,23 @@ const createContext = (app: AnyElysia, route: InternalRoute) => {
 
 	if (inference.query) fnLiteral += getQi
 
-	const getPath = !isDynamic
-		? `path:'${path}',`
-		: `get path(){` +
-			(inference.query ? '' : getQi) +
-			`if(qi===-1) return u.substring(s)\n` +
-			`return u.substring(s, qi)\n` +
-			`},`
+	const getPath = !inference.path
+		? ''
+		: !isDynamic
+			? `path:'${path}',`
+			: `get path(){` +
+				(inference.query ? '' : getQi) +
+				`if(qi===-1) return u.substring(s)\n` +
+				`return u.substring(s, qi)\n` +
+				`},`
 
 	fnLiteral +=
-		`const c={request,` +
+		allocateIf(`const c=`, !isInline) +
+		`{request,` +
 		`store,` +
 		allocateIf(`qi,`, inference.query) +
-		allocateIf(`route`, inference.route || hasTrace) +
+		allocateIf(`route,`, inference.route || hasTrace) +
+		`params:request.params,` +
 		getPath +
 		`url:request.url,` +
 		`redirect,` +
@@ -60,9 +70,7 @@ const createContext = (app: AnyElysia, route: InternalRoute) => {
 
 	fnLiteral += `,status:200}`
 
-	// @ts-expect-error private
-	if (app.inference.server)
-		fnLiteral += `,get server(){return app.getServer()}`
+	if (inference.server) fnLiteral += `,get server(){return app.getServer()}`
 
 	if (hasTrace) fnLiteral += ',[ELYSIA_REQUEST_ID]:id'
 
@@ -81,17 +89,39 @@ const createContext = (app: AnyElysia, route: InternalRoute) => {
 }
 
 export const createBunRouteHandler = (app: AnyElysia, route: InternalRoute) => {
+	const hasTrace = !!app.event.trace?.length
+	// @ts-expect-error private property
+	const hasHoc = !!app.extender.higherOrderFunctions.length
+
+	const inference = sucrose(
+		Object.assign({}, route.hooks, {
+			handler: route.handler
+		}),
+		// @ts-expect-error
+		app.inference
+	)
+
 	let fnLiteral =
 		'const handler=data.handler,' +
 		'store=data.store,' +
 		'redirect=data.redirect,' +
+		allocateIf('route=data.route,', inference.route || hasTrace) +
+		allocateIf('randomId=data.randomId,', hasTrace) +
+		allocateIf(`ELYSIA_REQUEST_ID=data.ELYSIA_REQUEST_ID,`, hasTrace) +
+		allocateIf(`ELYSIA_TRACE=data.ELYSIA_TRACE,`, hasTrace) +
+		allocateIf(`trace=data.trace,`, hasTrace) +
+		allocateIf(`hoc=data.hoc,`, hasHoc) +
 		'error=data.error\n' +
 		'function map(request){'
 
-	fnLiteral += createContext(app, route)
-	fnLiteral += createOnRequestHandler(app)
+	if (hasTrace || app.event.request?.length) {
+		fnLiteral += createContext(app, route, inference)
+		fnLiteral += createOnRequestHandler(app)
 
-	fnLiteral += 'return handler(c)}\n'
+		fnLiteral += 'return handler(c)}'
+	} else {
+		fnLiteral += `return handler(${createContext(app, route, inference, true)})}`
+	}
 
 	fnLiteral += createHoc(app)
 
@@ -102,6 +132,13 @@ export const createBunRouteHandler = (app: AnyElysia, route: InternalRoute) => {
 		handler: route.compile?.() ?? route.composed,
 		redirect,
 		error,
-		store: app.store
+		// @ts-expect-error private property
+		hoc: app.extender.higherOrderFunctions.map((x) => x.fn),
+		store: app.store,
+		route: inference.route || hasTrace ? route.path : undefined,
+		randomId: hasTrace ? randomId : undefined,
+		ELYSIA_TRACE: hasTrace ? ELYSIA_TRACE : undefined,
+		ELYSIA_REQUEST_ID: hasTrace ? ELYSIA_REQUEST_ID : undefined,
+		trace: hasTrace ? app.event.trace?.map((x) => x?.fn ?? x) : undefined
 	})
 }